refactoring Runner etc to lay the groundwork for --watch

+++ packages/element/src/Element.ts
@@ -1,20 +1,21 @@
 import { Logger } from 'winston'
-import PuppeteerDriver from './driver/Puppeteer'
 import { IReporter } from './Reporter'
+import { PuppeteerClient, launch } from './driver/Puppeteer'
 import { RuntimeEnvironment } from './runtime-environment/types'
-import { Browser } from './types'
 import Runner from './Runner'
 import { ITestScript, TestScriptOptions, mustCompileFile } from './TestScript'
 import { TestSettings } from './runtime/Settings'
 import { TestObserver } from './runtime/test-observers/Observer'
+import { AsyncFactory } from './utils/Factory'
 
 export interface ElementOptions {
 	logger: Logger
 	runEnv: RuntimeEnvironment
 	reporter: IReporter
+	clientFactory?: AsyncFactory<PuppeteerClient>
 	testScript: string
 	strictCompilation: boolean
-	driver?: { new (): Browser }
+	watch: boolean
 	headless: boolean
 	devtools: boolean
 	chrome: string | boolean
@@ -35,13 +36,11 @@ export function runUntilExit(fn: () => Promise<void>) {
 }
 
 export async function runCommandLine(opts: ElementOptions): Promise<void> {
-	let { logger, testScript, driver } = opts
-
-	if (!driver) driver = PuppeteerDriver
+	let { logger, testScript, clientFactory } = opts
 
 	const runner = new Runner(
+		clientFactory || launch,
 		opts.runEnv,
-		driver,
 		opts.reporter,
 		logger,
 		opts.testSettingOverrides,
@@ -54,17 +53,21 @@ export async function runCommandLine(opts: ElementOptions): Promise<void> {
 		opts.testObserverFactory,
 	)
 
-	process.on('SIGINT', async () => {
-		logger.debug('Received SIGINT')
-		await runner.shutdown()
-	})
+	const installSignalHandlers = true
 
-	process.once('SIGUSR2', async () => {
-		// Usually received by nodemon on file change
-		logger.debug('Received SIGUSR2')
-		await runner.shutdown()
-		process.kill(process.pid, 'SIGUSR2')
-	})
+	if (installSignalHandlers) {
+		process.on('SIGINT', async () => {
+			logger.debug('Received SIGINT')
+			await runner.stop()
+		})
+
+		process.once('SIGUSR2', async () => {
+			// Usually received by nodemon on file change
+			logger.debug('Received SIGUSR2')
+			await runner.stop()
+			process.kill(process.pid, 'SIGUSR2')
+		})
+	}
 
 	logger.debug(`Loading test script: ${testScript}`)
 
@@ -73,6 +76,8 @@ export async function runCommandLine(opts: ElementOptions): Promise<void> {
 		traceResolution: false,
 	}
 
-	const testScriptObj: ITestScript = await mustCompileFile(testScript, testScriptOptions)
-	await runner.run(testScriptObj)
+	const testScriptFactory = async (): Promise<ITestScript> => {
+		return await mustCompileFile(testScript, testScriptOptions)
+	}
+	await runner.run(testScriptFactory)
 }
